Improved some validation rules
(i) Made isValidParentId() more generic so that it can validate against any
model and any column name (with the default of column `id' in the current
model).

(ii) Renamed isUniqueFieldAmongstPeers() to isUniqueNodeAmongstPeers() to better
describe what it does.

(iii) Wrote isPositiveInteger as it turns out that using array('decimal', 0) as
a validation rule doesn't actually do what I want it to.  Whilst I could've used
array('custom', '/^\d+$/') as the rule, this uses preg_match which is quite
expensive in terms of performance, so I use a combination of is_int() and
ctype_digit().  If that doesn't quite make sense, then just look at lines
141-150 in www/app/app_model.php.