Fixed bug where failed validation would generate an edit form
The bug was that if validation failed on an add() form, then it wouldn't
redirect so $this->data would still present.  The Form helper would therefore
generate an edit-type form, and submit to /controller/edit (which would choke
because the row didn't already exist).

The fix is to explicitly define the form action (in the HTML attribute sense of
the word "action").