Re-indented PHP code to conform with our coding standards [#61 state:resolved]
The only characters that should have changed in this commit are tabs, spaces and
newlines (`git diff -w` is my friend).  Here's a more comprehensive list of
what's been done.

* Ensured whitespace at the beginning of lines is appropriate for that scope
* Corrected all indentation to use tabs, not spaces
* Made sure control structures and array declarations conform to coding
  standards on the wiki
* Removed extraneous empty lines and added a few in to space code
* Made sure all files end in a newline
* Checked that all files used LF as their line ending (they did)

Note that I haven't touched any of the views, as they're more HTML than PHP so
it's difficult to interpret the coding standards in this context.